Move TerrainState to its own file since its pub
