+++ src/generation/object/lib/mod.rs
@@ -1,14 +1,17 @@
 mod cell;
-mod connection_type;
+mod connection;
 mod object_data;
 mod object_grid;
 mod object_name;
+mod terrain_state;
 pub mod tile_data;
 mod wfc_status;
 
 pub use cell::Cell;
-pub use connection_type::Connection;
+pub use connection::Connection;
 pub use object_data::ObjectData;
 pub use object_grid::ObjectGrid;
 pub use object_name::ObjectName;
+pub use terrain_state::TerrainState;
+pub use tile_data::TileData;
 pub use wfc_status::IterationResult;
