+++ src/generation/object/lib/object_grid.rs
@@ -2,10 +2,8 @@ use crate::constants::CHUNK_SIZE;
 use crate::coords::Point;
 use crate::coords::point::{ChunkGrid, InternalGrid};
 use crate::generation::lib::{TerrainType, TileType};
-use crate::generation::object::lib::connection_type::get_connection_points;
-use crate::generation::object::lib::tile_data::TileData;
-use crate::generation::object::lib::{Cell, Connection, ObjectName};
-use crate::generation::resources::TerrainState;
+use crate::generation::object::lib::connection::get_connection_points;
+use crate::generation::object::lib::{Cell, Connection, ObjectName, TerrainState, TileData};
 use bevy::log::*;
 use bevy::platform::collections::HashMap;
 use bevy::reflect::Reflect;
